% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodetfakepkg5.R
\name{tinycodetfakepkg1}
\alias{tinycodetfakepkg1}
\alias{fun_paste}
\alias{\%paste0\%}
\title{tinycodetfakepkg1}
\usage{
fun_paste(x, y)

x \%paste0\% y
}
\arguments{
\item{x}{whatever}

\item{y}{whatever}
}
\description{
Bla
}
